move funcitons into files
